% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\docType{class}
\name{Encuesta}
\alias{Encuesta}
\title{Clase R6 Encuesta}
\format{
R6Class object
}
\description{
Hereda de `Graficar` y permite generar saldos de opinión y gráficos
combinados (opinión, conocimiento, Ns/Nc).
}
\examples{
\dontrun{
g <- Encuesta$new(
  diseno = encuesta_demo$muestra$diseno,
  diccionario = dicc,
  colores = colores,
  color_principal = "pink",
  tema = tema_morant()
)

g$saldos_opinion(
  sufijo_opinion = "opinion_pm",
  cat_ns_nc = "Ns/Nc",
  sufijo_conoce = "conoce_pm",
  cat_conoce = "Sí",
  actores = c("astiazaran", "delrio"),
  positivas = c("Muy buena", "Buena"),
  negativas = c("Mala", "Muy mala"),
  regular = "Regular"
)
}
Inicializa la clase Encuesta

}
\section{Super class}{
\code{\link[morantviz:Graficar]{morantviz::Graficar}} -> \code{Encuesta}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Encuesta-new}{\code{Encuesta$new()}}
\item \href{#method-Encuesta-saldos_opinion}{\code{Encuesta$saldos_opinion()}}
\item \href{#method-Encuesta-clone}{\code{Encuesta$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="morantviz" data-topic="Graficar" data-id="agregar_saldo"><a href='../../morantviz/html/Graficar.html#method-Graficar-agregar_saldo'><code>morantviz::Graficar$agregar_saldo()</code></a></span></li>
<li><span class="pkg-link" data-pkg="morantviz" data-topic="Graficar" data-id="cambiarSigno_media"><a href='../../morantviz/html/Graficar.html#method-Graficar-cambiarSigno_media'><code>morantviz::Graficar$cambiarSigno_media()</code></a></span></li>
<li><span class="pkg-link" data-pkg="morantviz" data-topic="Graficar" data-id="contar_variables"><a href='../../morantviz/html/Graficar.html#method-Graficar-contar_variables'><code>morantviz::Graficar$contar_variables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="morantviz" data-topic="Graficar" data-id="envolver_etiquetas"><a href='../../morantviz/html/Graficar.html#method-Graficar-envolver_etiquetas'><code>morantviz::Graficar$envolver_etiquetas()</code></a></span></li>
<li><span class="pkg-link" data-pkg="morantviz" data-topic="Graficar" data-id="etiquetar_regular"><a href='../../morantviz/html/Graficar.html#method-Graficar-etiquetar_regular'><code>morantviz::Graficar$etiquetar_regular()</code></a></span></li>
<li><span class="pkg-link" data-pkg="morantviz" data-topic="Graficar" data-id="filtrar_respuesta"><a href='../../morantviz/html/Graficar.html#method-Graficar-filtrar_respuesta'><code>morantviz::Graficar$filtrar_respuesta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="morantviz" data-topic="Graficar" data-id="graficar_barras_divergente"><a href='../../morantviz/html/Graficar.html#method-Graficar-graficar_barras_divergente'><code>morantviz::Graficar$graficar_barras_divergente()</code></a></span></li>
<li><span class="pkg-link" data-pkg="morantviz" data-topic="Graficar" data-id="graficar_barras_h"><a href='../../morantviz/html/Graficar.html#method-Graficar-graficar_barras_h'><code>morantviz::Graficar$graficar_barras_h()</code></a></span></li>
<li><span class="pkg-link" data-pkg="morantviz" data-topic="Graficar" data-id="partir_regular"><a href='../../morantviz/html/Graficar.html#method-Graficar-partir_regular'><code>morantviz::Graficar$partir_regular()</code></a></span></li>
<li><span class="pkg-link" data-pkg="morantviz" data-topic="Graficar" data-id="pegar_color"><a href='../../morantviz/html/Graficar.html#method-Graficar-pegar_color'><code>morantviz::Graficar$pegar_color()</code></a></span></li>
<li><span class="pkg-link" data-pkg="morantviz" data-topic="Graficar" data-id="pegar_diccionario"><a href='../../morantviz/html/Graficar.html#method-Graficar-pegar_diccionario'><code>morantviz::Graficar$pegar_diccionario()</code></a></span></li>
<li><span class="pkg-link" data-pkg="morantviz" data-topic="Graficar" data-id="reordenar_columna"><a href='../../morantviz/html/Graficar.html#method-Graficar-reordenar_columna'><code>morantviz::Graficar$reordenar_columna()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Encuesta-new"></a>}}
\if{latex}{\out{\hypertarget{method-Encuesta-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Encuesta$new(diseno, diccionario, colores, color_principal, tema)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Encuesta-saldos_opinion"></a>}}
\if{latex}{\out{\hypertarget{method-Encuesta-saldos_opinion}{}}}
\subsection{Method \code{saldos_opinion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Encuesta$saldos_opinion(
  sufijo_opinion,
  cat_ns_nc,
  sufijo_conoce,
  cat_conoce,
  actores,
  positivas,
  negativas,
  regular
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Encuesta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Encuesta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Encuesta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
