% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\docType{class}
\name{Graficar}
\alias{Graficar}
\title{Clase R6 Graficar}
\format{
R6Class object
}
\description{
Clase base para construir tablas, manipular respuestas y generar
gráficas a partir de encuestas con diseño muestral y diccionario.
}
\examples{
\dontrun{
g <- Graficar$new(
  diseno = encuesta_demo$muestra$diseno,
  diccionario = dicc,
  colores = colores,
  color_principal = "pink",
  tema = tema_morant()
)

# Conocimiento en barras horizontales
g$
  contar_variables(
    variables = c("conoce_pm_astiazaran", "conoce_pm_delrio"),
    confint = FALSE
  )$
  filtrar_respuesta(valor = "Sí")$
  pegar_diccionario()$
  pegar_color()$
  envolver_etiquetas(columna = "nombre", ancho = 13)$
  reordenar_columna(columna = "nombre", tipo = "asc")

g$graficar_barras_h(x = "nombre")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tbl}}{Tibble con los resultados procesados.}

\item{\code{grafica}}{Objeto `ggplot` generado.}

\item{\code{diseno}}{Objeto survey design (de `survey`).}

\item{\code{diccionario}}{Diccionario de variables y etiquetas.}

\item{\code{colores}}{Tabla de colores por respuesta.}

\item{\code{color_principal}}{Color por defecto.}

\item{\code{tema}}{Tema de `ggplot` a aplicar en las gráficas.
Inicializar objeto Graficar}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Graficar-new}{\code{Graficar$new()}}
\item \href{#method-Graficar-contar_variables}{\code{Graficar$contar_variables()}}
\item \href{#method-Graficar-filtrar_respuesta}{\code{Graficar$filtrar_respuesta()}}
\item \href{#method-Graficar-reordenar_columna}{\code{Graficar$reordenar_columna()}}
\item \href{#method-Graficar-envolver_etiquetas}{\code{Graficar$envolver_etiquetas()}}
\item \href{#method-Graficar-pegar_diccionario}{\code{Graficar$pegar_diccionario()}}
\item \href{#method-Graficar-partir_regular}{\code{Graficar$partir_regular()}}
\item \href{#method-Graficar-cambiarSigno_media}{\code{Graficar$cambiarSigno_media()}}
\item \href{#method-Graficar-etiquetar_regular}{\code{Graficar$etiquetar_regular()}}
\item \href{#method-Graficar-pegar_color}{\code{Graficar$pegar_color()}}
\item \href{#method-Graficar-agregar_saldo}{\code{Graficar$agregar_saldo()}}
\item \href{#method-Graficar-graficar_barras_h}{\code{Graficar$graficar_barras_h()}}
\item \href{#method-Graficar-graficar_barras_divergente}{\code{Graficar$graficar_barras_divergente()}}
\item \href{#method-Graficar-clone}{\code{Graficar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficar-new"></a>}}
\if{latex}{\out{\hypertarget{method-Graficar-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficar$new(diseno, diccionario, colores, color_principal, tema)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficar-contar_variables"></a>}}
\if{latex}{\out{\hypertarget{method-Graficar-contar_variables}{}}}
\subsection{Method \code{contar_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficar$contar_variables(variables, confint)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficar-filtrar_respuesta"></a>}}
\if{latex}{\out{\hypertarget{method-Graficar-filtrar_respuesta}{}}}
\subsection{Method \code{filtrar_respuesta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficar$filtrar_respuesta(variable, valor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficar-reordenar_columna"></a>}}
\if{latex}{\out{\hypertarget{method-Graficar-reordenar_columna}{}}}
\subsection{Method \code{reordenar_columna()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficar$reordenar_columna(columna, tipo, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficar-envolver_etiquetas"></a>}}
\if{latex}{\out{\hypertarget{method-Graficar-envolver_etiquetas}{}}}
\subsection{Method \code{envolver_etiquetas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficar$envolver_etiquetas(columna, ancho)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficar-pegar_diccionario"></a>}}
\if{latex}{\out{\hypertarget{method-Graficar-pegar_diccionario}{}}}
\subsection{Method \code{pegar_diccionario()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficar$pegar_diccionario()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficar-partir_regular"></a>}}
\if{latex}{\out{\hypertarget{method-Graficar-partir_regular}{}}}
\subsection{Method \code{partir_regular()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficar$partir_regular(opcion)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficar-cambiarSigno_media"></a>}}
\if{latex}{\out{\hypertarget{method-Graficar-cambiarSigno_media}{}}}
\subsection{Method \code{cambiarSigno_media()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficar$cambiarSigno_media(negativo)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficar-etiquetar_regular"></a>}}
\if{latex}{\out{\hypertarget{method-Graficar-etiquetar_regular}{}}}
\subsection{Method \code{etiquetar_regular()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficar$etiquetar_regular(regular)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficar-pegar_color"></a>}}
\if{latex}{\out{\hypertarget{method-Graficar-pegar_color}{}}}
\subsection{Method \code{pegar_color()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficar$pegar_color()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficar-agregar_saldo"></a>}}
\if{latex}{\out{\hypertarget{method-Graficar-agregar_saldo}{}}}
\subsection{Method \code{agregar_saldo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficar$agregar_saldo(por)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficar-graficar_barras_h"></a>}}
\if{latex}{\out{\hypertarget{method-Graficar-graficar_barras_h}{}}}
\subsection{Method \code{graficar_barras_h()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficar$graficar_barras_h(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficar-graficar_barras_divergente"></a>}}
\if{latex}{\out{\hypertarget{method-Graficar-graficar_barras_divergente}{}}}
\subsection{Method \code{graficar_barras_divergente()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficar$graficar_barras_divergente(regular, positivas, negativas)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Graficar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
